requirements_WINDOWS.txt
altgraph==0.17.3
appdirs==1.4.4
asciitree==0.3.3
astroid==2.12.12
asttokens==2.0.8
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
darkdetect==0.7.1
dask==2022.9.2
debugpy==1.6.3
decorator==5.1.1
diffpy.structure==3.0.2
diffsims==0.5.0
dill==0.3.5.1
docstring-to-markdown==0.10
Eel==0.14.0
entrypoints==0.4
executing==1.1.0
fasteners==0.18
flake8==5.0.4
fonttools==4.37.4
fsspec==2022.8.2
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h5py==3.7.0
hyperspy==1.7.3
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
ipykernel==6.16.0
ipyparallel==8.4.1
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kikuchipy==0.8.4
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
llvmlite==0.39.1
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
mccabe==0.7.0
mpmath==1.2.1
natsort==8.2.0
nest-asyncio==1.5.6
networkx==2.8.7
nlopt==2.7.1
numba==0.56.2
numcodecs==0.10.2
numexpr==2.8.3
numpy==1.23.3
numpy-quaternion==2022.4.2
orix==0.11.0
packaging==21.3
parso==0.8.3
partd==1.3.0
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.19
platformdirs==2.6.2
pluggy==1.0.0
pooch==1.6.0
prettytable==3.4.1
prompt-toolkit==3.0.32
psutil==5.9.2
pure-eval==0.2.2
PyCifRW==4.4.4
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
pyebsdindex==0.1.1
pyflakes==2.5.0
Pygments==2.13.0
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2022.13
pylint==2.15.5
pyopencl==2022.3.1
pyparsing==3.0.9
pyqtdarktheme==2.1.0
pyqtgraph==0.13.1
PySide6==6.3.2
PySide6-Addons==6.3.2
PySide6-Essentials==6.3.2
pyswarms==1.3.0
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.6.0
pytoolconfig==1.2.2
pytools==2022.1.14
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
QtPy==2.2.1
requests==2.28.1
rope==1.4.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
shiboken6==6.3.2
six==1.16.0
snowballstemmer==2.2.0
sparse==0.13.0
stack-data==0.5.1
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
traits==6.4.1
transforms3d==0.4.1
typing_extensions==4.3.0
ujson==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
whatthepatch==1.0.3
whichcraft==0.6.1
wrapt==1.14.1
yapf==0.32.0
zarr==2.13.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.5.2